Fixed `隐藏了导航栏的根控制器`pop时, 视图层级的问题

+++ SJVideoPlayerBackGR/UINavigationController+SJVideoPlayerAdd.m
@@ -389,7 +389,7 @@ static __weak UIViewController *_tmpShowViewController;
         if ( ![self.view isDescendantOfView:obj] ) return ;
         *stop = YES;
         // Move the `screenshot` to the bottom of the `current view`.
-        [window insertSubview:self.SJ_screenshotView atIndex:idx - 1];
+        [window insertSubview:self.SJ_screenshotView atIndex:(int)idx - 1];
     }];
 
     self.SJ_screenshotView.hidden = NO;
